Organise ActionForm stories
